add UserPanelContainer
