+++ src/pages/TestEditingPage.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import TestEditingContainer from 'containers/TestEditing/TestEditingContainer';
 import withUserAuthentication from 'components/hoc/withUserAuthentication';
-import UserPanel from 'components/UserPanel/UserPanel';
+import UserPanelContainer from 'containers/UserPanelContainer';
 
 function TestEditingPage() {
   return (
     <>
-      <UserPanel />
+      <UserPanelContainer />
       <TestEditingContainer />
     </>
   );
